QC Pass (#39)
* update

* small fixes

* small fixes

* more refactors

* more qc fixes

* more correction passes

* path cleanups

* ore path fixes

* some Bevy 0.15 fixes

+++ ferritin-core/src/conversions/pse.rs
@@ -52,7 +52,7 @@ impl From<&PSEData> for AtomCollection {
             .multiunzip();
 
         AtomCollection::new(
-            atoms.len() as usize, // size
+            atoms.len(),
             coords,
             res_ids,
             res_names,
